reorg attributes, fix minor stuff, write compile-time tests

+++ src/main/scala/ohnosequences/saws/dynamodb/package.scala
@@ -13,7 +13,7 @@ package object dynamodb {
 
   // dynamodb type restrictions
   type Bytes = Seq[Byte]
-  type Num = Long
+  type Num   = Int
   type Values = either[Num]#or[String]#or[Bytes]#or[Set[Num]]#or[Set[String]]#or[Set[Bytes]]
   // not documented
   type PrimaryKeyValues = either[String]#or[Num]
